+++ src/config.cpp
@@ -1,3 +1,3 @@
 #include "config.hpp"
 
-const Pair<uint16_t, uint16_t> Config::BALL_RATE = {1, 1};
+const Pair<uint8_t, uint8_t> Config::BALL_RATE = {1, 1};
